赖好能动

+++ src/px_user_func.h
@@ -1,7 +1,7 @@
 #ifndef _PX_USER_FUNC_H_
 #define _PX_USER_FUNC_H_
 
-int px_user_init(struct px_gui *gui, int width, int height);
+uint8_t px_user_init(struct px_gui *gui, int width, int height);
 void px_user_deinit(struct px_gui *gui);
 
 void px_user_render(struct px_gui *gui);
@@ -10,3 +10,4 @@ void px_user_poll_event(struct px_gui *gui);
 void px_delay_at_least(struct px_gui *gui, int ms);
 
 #endif
+
